[FIX] Fixed avg time calculation.

+++ src/ar_Nccl.cu
@@ -453,7 +453,7 @@ int main(int argc, char *argv[])
     for(int j=fix_buff_size; j<max_j; j++){
         long int N = 1 << j;
         long int B_in_GB = 1 << 30;
-        long int num_B = sizeof(dtype)*N*((size-1)/size)*2;
+        long int num_B = sizeof(dtype)*N*((size-1)/(float)size)*2;
         double num_GB = (double)num_B / (double)B_in_GB;
 
         double avg_time_per_transfer = 0.0;
